Change API calls from depreciated getUsername()
I've replaced all of the client.getUsername() calls with getLocalPlayer.getName(). This fixes the issue that exists when using the Jagex Launcher.

You may need to clear out/rename some of the json entries from the blank file name. Also, this returns a case-sensitive username, but I'm not sure if previously getUsername() returned all lowercase or not.